moved updateFilesNode from editor to fileManager
