+++ frontend/src/components/EditorPage/FileManager/index.jsx
@@ -14,6 +14,7 @@ import { findNodeById, findNodeByName, getParentId, getPath, isFolder, removeByI
 import CustomDragHandle from '../../CustomDragHandle';
 import { Downgraded } from '@hookstate/core';
 import { useGetFilesNode } from '../../../graphql/getFilesNode';
+import { useUpdateFilesNode } from '../../../graphql/updateFilesNode';
 
 const MOCK_ROOT_ID = 'all';
 
@@ -27,7 +28,7 @@ const FileManagerColumn = forwardRef(({ children, ...rest }, ref) => {
     const [workerGroups, setWorkerGroups] = useState([]);
     const [selected, setSelected] = useState(null);
     const [expanded, setExpanded] = useState([]);
-    // const [data, setData] = useState(MOCK_DATA);
+    const [data, setData] = useState([]);
 
     // File/folder creation states
     const [isAddingNew, setIsAddingNew] = useState(false);
@@ -42,7 +43,8 @@ const FileManagerColumn = forwardRef(({ children, ...rest }, ref) => {
     const editingFileRef = useRef();
 
     // Graphql hook
-    const getFilesNode = useGetFilesNodeHook(rest.pipeline, rest.setData);
+    const getFilesNode = useGetFilesNodeHook(rest.pipeline, setData);
+    const updateFilesNode = useUpdateFilesNodeHook(rest.pipeline.environmentID, rest.pipeline.pipelineID, rest.pipeline.nodeID);
 
     // Check if selected file changed
     useEffect(() => {
@@ -132,7 +134,7 @@ const FileManagerColumn = forwardRef(({ children, ...rest }, ref) => {
         // Check if enter has being pressed
         if (e.charCode === 13) {
             const check = checkFileName(newFileName);
-            const alreadyExistsFile = findNodeByName(rest.data?.children, newFileName);
+            const alreadyExistsFile = findNodeByName(data?.children, newFileName);
 
             if (!check) return;
             if (alreadyExistsFile) {
@@ -159,9 +161,9 @@ const FileManagerColumn = forwardRef(({ children, ...rest }, ref) => {
 
         // Push new file to root children
         //Find the ID first
-        const currentSelectedElement = findNodeById(rest.data.children, selected); //
-        if (selected === rest.data.id) {
-            rest.data.children.push(newFileMock);
+        const currentSelectedElement = findNodeById(data.children, selected); //
+        if (selected === data.id) {
+            data.children.push(newFileMock);
         } else if (currentSelectedElement && currentSelectedElement.children) {
             currentSelectedElement.children.push(newFileMock);
             selectAndOpenNewFile(newFileMock);
@@ -198,7 +200,7 @@ const FileManagerColumn = forwardRef(({ children, ...rest }, ref) => {
         // Check if enter has being pressed
         if (e.charCode === 13) {
             const check = checkFolderName(newFolderName);
-            const alreadyExistsFolder = findNodeByName(rest.data?.children, newFolderName);
+            const alreadyExistsFolder = findNodeByName(data?.children, newFolderName);
 
             if (!check) return;
             if (alreadyExistsFolder) {
@@ -222,9 +224,9 @@ const FileManagerColumn = forwardRef(({ children, ...rest }, ref) => {
 
         // Push new file to root children
         //Find the ID first
-        const currentSelectedElement = findNodeById(rest.data.children, selected);
-        if (selected === rest.data.id) {
-            rest.data.children.push(newFolderMock);
+        const currentSelectedElement = findNodeById(data.children, selected);
+        if (selected === data.id) {
+            data.children.push(newFolderMock);
         } else if (currentSelectedElement && currentSelectedElement.children) {
             currentSelectedElement.children.push(newFolderMock);
             selectAndOpenNewFile(newFolderMock);
@@ -254,12 +256,12 @@ const FileManagerColumn = forwardRef(({ children, ...rest }, ref) => {
 
     const handleEditKeyPress = (e, nodes) => {
         if (e.charCode === 13) {
-            const folder = isFolder(selected, rest.data);
-            const elementToChange = findNodeById(rest.data.children, selected);
+            const folder = isFolder(selected, data);
+            const elementToChange = findNodeById(data.children, selected);
 
             if (folder) {
                 const check = checkFolderName(tmpFileName);
-                const alreadyExistsFolder = findNodeByName(rest.data?.children, tmpFileName);
+                const alreadyExistsFolder = findNodeByName(data?.children, tmpFileName);
 
                 if (!check) {
                     setIsEditing(false);
@@ -283,7 +285,7 @@ const FileManagerColumn = forwardRef(({ children, ...rest }, ref) => {
                 }
             } else {
                 const check = checkFileName(tmpFileName);
-                const alreadyExistsFile = findNodeByName(rest.data?.children, tmpFileName);
+                const alreadyExistsFile = findNodeByName(data?.children, tmpFileName);
 
                 if (!check) {
                     setIsEditing(false);
@@ -313,9 +315,9 @@ const FileManagerColumn = forwardRef(({ children, ...rest }, ref) => {
     // ####### Delete ########
     const handleDeleteIconClick = () => {
         // Check if it's file or folder
-        const isTryingToDeleteFolder = isFolder(selected, rest.data);
+        const isTryingToDeleteFolder = isFolder(selected, data);
 
-        const current = findNodeById(rest.data.children, selected);
+        const current = findNodeById(data.children, selected);
         let message = `Are you sure you want to delete - ${current?.name}?`;
 
         if (current?.children && current?.children.length > 0) {
@@ -325,9 +327,9 @@ const FileManagerColumn = forwardRef(({ children, ...rest }, ref) => {
         const askForConfirmation = window.confirm(message);
 
         if (askForConfirmation === true) {
-            const removed = removeById(rest.data.children, selected);
-            const newData = { ...rest.data, children: removed };
-            rest.setData(newData);
+            const removed = removeById(data.children, selected);
+            const newData = { ...data, children: removed };
+            setData(newData);
 
             // Remove child from tabs
             if (current?.children && current?.children.length > 0) {
@@ -351,12 +353,12 @@ const FileManagerColumn = forwardRef(({ children, ...rest }, ref) => {
     // ####### Utils ########
     // Checks if selected folder is expanded
     const expandFolder = () => {
-        if (!selected || selected === rest.data.id) {
-            if (expanded?.filter((id) => id === rest.data.id).length > 0) {
+        if (!selected || selected === data.id) {
+            if (expanded?.filter((id) => id === data.id).length > 0) {
                 return;
             } else {
-                setExpanded([...expanded, rest.data.id]);
-                setSelected(rest.data.id);
+                setExpanded([...expanded, data.id]);
+                setSelected(data.id);
             }
         } else {
             if (expanded?.filter((id) => id === selected).length > 0) {
@@ -369,16 +371,16 @@ const FileManagerColumn = forwardRef(({ children, ...rest }, ref) => {
 
     // Check if it is folder of file
     const checkLevel = () => {
-        if (!selected || selected === rest.data.id) return;
+        if (!selected || selected === data.id) return;
 
-        const selectedInfo = findNodeById(rest.data.children, selected);
+        const selectedInfo = findNodeById(data.children, selected);
         if (selectedInfo) {
             if (selectedInfo.children) {
                 // Is folder\
                 return false;
             } else {
                 // Is file
-                const closestParentID = getParentId(rest.data.children, selected);
+                const closestParentID = getParentId(data.children, selected);
                 closestParentID && setSelected(closestParentID);
             }
         }
@@ -495,7 +497,7 @@ const FileManagerColumn = forwardRef(({ children, ...rest }, ref) => {
                     />
                 }
                 onClick={() => handleFileClick(nodes)}>
-                <Box className={`${nodes.id === rest.data.id ? 'showOnHover' : ''} hidden_controls tree-${nodes.id}`}>
+                <Box className={`${nodes.id === data.id ? 'showOnHover' : ''} hidden_controls tree-${nodes.id}`}>
                     <IconButton aria-label="Edit File" onClick={handleEdit}>
                         <Box component={FontAwesomeIcon} icon={faPencilAlt} sx={{ color: 'editorPage.fileManagerIcon', fontSize: 9 }} />
                     </IconButton>
@@ -547,7 +549,7 @@ const FileManagerColumn = forwardRef(({ children, ...rest }, ref) => {
         Editor.tabs.set((prevTabs) => [...prevTabs, file]);
 
         // Set file path
-        const path = getPath(rest.data.children, selected);
+        const path = getPath(data.children, selected);
         Editor.currentPath.set(path);
     };
 
@@ -604,7 +606,7 @@ const FileManagerColumn = forwardRef(({ children, ...rest }, ref) => {
 
                         p: '0 5px',
                     }}>
-                    {renderTree(rest.data)}
+                    {renderTree(data)}
                 </TreeView>
             </Box>
 
@@ -777,4 +779,64 @@ function prepareForFrontEnd(data) {
 
     recursive2(top.children);
     return top;
+}
+
+// ----- Custom hook
+export const useUpdateFilesNodeHook = (environmentID, pipelineID, nodeID) => {
+    // GraphQL hook
+    const updateFilesNode = useUpdateFilesNode();
+
+    const { enqueueSnackbar, closeSnackbar } = useSnackbar();
+
+    // Get files
+    return async (data) => {
+        const input = prepareFilesNodeForBackend(data, environmentID, pipelineID, nodeID);
+        // return;
+        const response = await updateFilesNode({ input });
+
+        if (response.r === 'error') {
+            closeSnackbar();
+            enqueueSnackbar("Can't get files: " + response.msg, { variant: 'error' });
+        } else if (response.errors) {
+            response.errors.map((err) => enqueueSnackbar(err.message, { variant: 'error' }));
+        } else {
+            enqueueSnackbar('Success', { variant: 'success' });
+        }
+    };
+};
+
+function prepareFilesNodeForBackend(files, environmentID, pipelineID, nodeID) {
+    let array = [];
+    array.push({
+        folderID: files.id,
+        parentID: files.parentID,
+        environmentID,
+        pipelineID,
+        nodeID,
+        folderName: files.name,
+        fType: files.fType,
+        active: true,
+    });
+
+    function recursive(arr) {
+        for (const key of arr) {
+            if (key.children) {
+                recursive(key.children);
+            }
+            array.push({
+                folderID: key.id,
+                parentID: key.parentID,
+                environmentID,
+                pipelineID,
+                nodeID,
+                folderName: key.name,
+                fType: key.fType,
+                active: true,
+            });
+        }
+        return array;
+    }
+    recursive(files.children);
+
+    return array;
 }
